requirements.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.2
flake8==4.0.1
idna==3.3
ipython==8.0.0
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.3
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
